refactor: Split code in Mail-Core
